[CREATE] : Create Course Category response body data model and refactor previous data model name
The commit introduces a new data model, `CategoryListModel`, to represent course category data within the `CourseCategoriesResponseBody`. It also refactors the name of the previous data model from `Data` to `CategoryListModel` for better clarity. The change is done by creating the `CategoryListModel` class with a factory constructor. Also it update the json parsing in the class `CourseCategoriesResponseBody`.